+++ CHANGELOG.md
@@ -10,6 +10,7 @@
 - (Improvement) Change default logging level to info. Add --log.sampling (default true). Adjust log levels.
 - (Maintenance) Bump Go to 1.21.6
 - (Bugfix) Enable LazyLoader for CRD & CRD Schemas
+- (Feature) (ML) Restore ReadinessProbe for ML Storage sidecar
 
 ## [1.2.36](https://github.com/arangodb/kube-arangodb/tree/1.2.36) (2024-01-08)
 - (Documentation) Improvements and fixes for rendered documentation (GH pages)